fix(messaging): fixed the messaging search and re-orged the files (#319)
* fix(messaging): fixed the messaging search and re-orged the files

* chore(snapshote): fixed snapshot for messaging list
