+++ components/organisms/messaging/background/README.md
@@ -1,3 +1,4 @@
-# Messaging/MessageBackground
+# Messaging/Background
 
-Provides a background for the messages, we have 3 examples here: Whatsapp, drykiss and blank background
+Provides a background for the messages, we have 3 examples here: Whatsapp, drykiss and blank
+background.
