--- components/organisms/messaging/background/components/index.js
@@ -1,1 +1,0 @@
-export * from './messageBackground'
