--- components/organisms/messaging/background/index.js
@@ -1,1 +1,0 @@
-export * from './components'
