+++ components/organisms/messaging/base/components/icon.js
@@ -1,5 +1,5 @@
 /**
- * Messaging/Message/Icon
+ * Messaging/Base/Icon
  */
 
 // React
