+++ components/organisms/messaging/base/components/index.js
@@ -1,0 +1,4 @@
+/**
+ * Messaging/Base
+ */
+export * from './base'
