+++ components/organisms/messaging/base/index.js
@@ -1,4 +1,4 @@
 /**
- * Messaging/Send
+ * Messaging/Base
  */
 export * from './components'
