+++ components/organisms/messaging/container/README.md
@@ -1,5 +1,4 @@
 # Messaging/Container
 
-Provides an message container where we can find different examples of messages
-
-Feature also include messages being truncated by Height
+Provides an message container where we can find different examples of messages.
+Feature also include messages being truncated by Height.
