--- components/organisms/messaging/container/components/index.js
@@ -1,1 +1,0 @@
-export * from './container'
