--- components/organisms/messaging/container/index.js
@@ -1,1 +1,0 @@
-export * from './components'
