+++ components/organisms/messaging/list/README.md
@@ -1,1 +1,1 @@
-# Messaging/MessageList
+# Messaging/List
