+++ components/organisms/messaging/list/__mocks__/list.js
@@ -1,5 +1,5 @@
 /**
- * Messages data
+ * Messaging/List
  */
 export const messages = [
   {
