+++ components/organisms/messaging/list/__tests__/stories.js
@@ -1,5 +1,5 @@
 /**
- * Messaging/MessageList
+ * Messaging/List
  */
 
 // React
@@ -9,14 +9,14 @@ import React from 'react'
 import { Wrapper } from 'decorators'
 
 // UI
-import { MessageList } from '../'
+import { MessageList } from 'components'
 import Readme from '../README.md'
 
 // Data
 import { messages } from '../__mocks__/list'
 
 export default {
-  title: 'Organisms/Messaging/MessageList',
+  title: 'Organisms/Messaging/List',
   component: MessageList,
   decorators: [Wrapper],
   parameters: {
