--- components/organisms/messaging/list/components/index.js
@@ -1,1 +1,0 @@
-export * from './messageList'
