+++ components/organisms/messaging/list/list.js
@@ -1,5 +1,5 @@
 /**
- * Messages
+ * Messaging/List
  */
 
 // React
@@ -7,7 +7,7 @@ import React from 'react'
 import { array } from 'prop-types'
 
 // UI
-import { Message } from '../../'
+import { Message } from '../../../'
 
 export const MessageList = ({ messages }) => {
   return messages.map((message, index) => (
