--- components/organisms/messaging/message/components/index.js
@@ -1,4 +1,0 @@
-/**
- * Messaging - Message
- */
-export * from './message'
