--- components/organisms/messaging/messageBase/index.js
@@ -1,4 +1,0 @@
-/**
- * Messaging/Message
- */
-export * from './components'
