--- components/organisms/messaging/search/components/index.js
@@ -1,1 +1,0 @@
-export * from './search'
