--- components/organisms/messaging/search/index.js
@@ -1,4 +1,0 @@
-/**
- * Messaging/Search
- */
-export * from './components'
