+++ components/organisms/messaging/send/README.md
@@ -2,9 +2,10 @@
 
 Provides a messaging send box, attach, emoji and send button.
 
-- @features
-- - Dropdown to select audience to send too.
-- - Attach file
-- - Emoji selection
-- - Send button
-- - Write message input
+@features
+
+- Dropdown to select audience to send too.
+- Attach file
+- Emoji selection
+- Send button
+- Write message input
