--- components/organisms/messaging/send/components/index.js
@@ -1,4 +1,0 @@
-/**
- * Messaging/Send
- */
-export * from './send'
