+++ components/organisms/messaging/tail/tail.js
@@ -1,5 +1,5 @@
 /**
- * MessageTail
+ * Messaging/Tail
  */
 
 // React
